+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - MOB-Suite
 - gMATS
 - MenDeVar
+- Azure service bus to handle the insertion of isolates in BIGSdb
 
 ### Changed
 - All Jammy existing assays to bigsdb for neisseria listeria salmonella mycobacterium influenza.
@@ -27,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed:
 - PointFinder, ResFinder, NCBI AMR, PlasmidFinder
 - Ability to support stec species
+- cron job to target insertion from MongoDB Atlas to BIGSdb
 
 ## [2.1.1 - support/2.1.0]
 ### Changed: